fix(Editor): separate close and disconnect functions
* `close` is for closing the editor.
  It tries to save the document and clean everything up.
* `disconnect` is for cleaning up the current collaboration sessions.
  It will not save the document and asumes the editing will be resumed
  after a reconnect.

Move `sendRemainingSteps` out to the sync service.
Also make close in the websocket polyfill sync.
Just clean up the polyfills state.
