+++ src/services/WebSocketPolyfill.js
@@ -97,36 +97,11 @@ export default function initWebSocketPolyfill(syncService, fileId, initialSessio
 		}
 
 		async close() {
-			await this.#sendRemainingSteps()
 			Object.entries(this.#handlers)
 				.forEach(([key, value]) => syncService.off(key, value))
 			this.#handlers = []
-			syncService.close().then(() => {
-				this.onclose()
-			})
+			this.onclose()
 			logger.debug('Websocket closed')
-		}
-
-		#sendRemainingSteps() {
-			if (queue.length) {
-				let outbox = []
-				return syncService.sendStepsNow(() => {
-					const data = {
-						steps: this.#steps,
-						awareness: this.#awareness,
-						version: this.#version,
-					}
-					outbox = [...queue]
-					logger.debug('sending final steps ', data)
-					return data
-				})?.then(() => {
-					// only keep the steps that were not send yet
-					queue.splice(0,
-						queue.length,
-						...queue.filter(s => !outbox.includes(s)),
-					)
-				}, err => logger.error(err))
-			}
 		}
 
 	}
